requirements.txt
python-telegram-bot==20.5
requests==2.31.0
python-dotenv==1.0.0
pillow==10.0.0
base64io==1.0.3
SpeechRecognition==3.10.0
pydub==0.25.1
ffmpeg-python==0.2.0
openai-whisper==20231117
soundfile==0.12.1 